pyproject.toml
[project]
authors = [{name = "AmadFat", email = "1943269060@qq.com"}]
dependencies = [
    "torch",
    "torchvision",
    "tabulate",
    "tensorboard",
    "alive-progress",
    "lmdb",
    "pillow",
    "line-profiler",
]
name = "nnaf"
requires-python = ">=3.12,<3.13"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]
